refactor: moved my schema defining object files
