+++ IndoorMapsBackend/prisma/schema.prisma
@@ -25,7 +25,7 @@ model Floor {
   description String
   building    Building @relation(fields: [buildingId], references: [id], onDelete: Cascade, onUpdate: Cascade)
   buildingId  Int
-  shape       String?
+  shape       Json?
 }
 
 model Building {
@@ -42,7 +42,7 @@ model Area {
   id          Int     @id @default(autoincrement())
   title       String
   description String
-  shape       String
+  shape       Json
   traversable Boolean
   category    String
   floor       Floor   @relation(fields: [floorId], references: [id], onDelete: Cascade, onUpdate: Cascade)
