+++ IndoorMapsBackend/src/auth/validateUser.ts
@@ -1,5 +1,5 @@
 import { isInstance } from "class-validator";
-import { User } from "../User.js";
+import { User } from "../graphqlSchemaTypes/User.js";
 import { Context } from "../context.js";
 import { verifyAccessToken } from "./jwt.js";
 import { GraphQLError } from "graphql";
