+++ IndoorMapsBackend/src/graphqlSchemaTypes/Area.ts
@@ -1,0 +1,37 @@
+import 'reflect-metadata'
+import { ObjectType, Field, ID, Float, Int } from 'type-graphql'
+
+import { MutationResult } from '../utils/generic.js'
+
+@ObjectType()
+export class NewAreaResult extends MutationResult {
+  @Field(type => Int)
+  databaseId: number
+
+  @Field(type => Int)
+  floorDatabaseId: number
+}
+
+@ObjectType()
+export class Area {
+  @Field((type) => ID)
+  id: string
+
+  @Field(type => Int)
+  databaseId: number
+
+  @Field()
+  title: string
+
+  @Field()
+  description: string
+
+  @Field()
+  shape: string
+
+  @Field()
+  traversable: boolean
+
+  @Field()
+  category: string
+}
