+++ IndoorMapsBackend/src/graphqlSchemaTypes/Floor.ts
@@ -1,0 +1,31 @@
+import { Field, ID, Int, ObjectType } from "type-graphql"
+
+import { MutationResult } from '../utils/generic.js'
+
+
+@ObjectType()
+export class NewFloorResult extends MutationResult {
+  @Field(type => Int)
+  databaseId: number
+
+  @Field(type => Int)
+  buildingDatabaseId: number
+}
+
+@ObjectType()
+export class Floor {
+  @Field((type) => ID)
+  id: string
+
+  @Field(type => Int)
+  databaseId: number
+
+  @Field()
+  title: string
+
+  @Field()
+  description: string
+
+  @Field({ nullable: true })
+  shape?: string
+}
