+++ IndoorMapsBackend/src/graphqlSchemaTypes/User.ts
@@ -3,7 +3,7 @@ import { ObjectType, Field, ID, registerEnumType } from 'type-graphql'
 import { IsEmail } from 'class-validator'
 
 import { Building } from './Building.js';
-import { MutationResult } from './utils/generic.js';
+import { MutationResult } from '../utils/generic.js';
 
 @ObjectType()
 export class User {
