+++ IndoorMapsBackend/src/utils/typeConversions.ts
@@ -2,9 +2,11 @@
  * this file containes functions that convert between the database type and the Graphql Types, It is used any time data is sent from the database to the front end
 */
 
-import { User } from '../User.js'
-import { Area, Building, Floor, LatLng } from '../Building.js'
+import { User } from '../graphqlSchemaTypes/User.js'
+import { Building, LatLng } from '../graphqlSchemaTypes/Building.js'
 import { Building as DbBuilding, Floor as DbFloor, Area as DbArea, User as DbUser } from '@prisma/client'
+import { Floor } from '../graphqlSchemaTypes/Floor.js'
+import { Area } from '../graphqlSchemaTypes/Area.js'
 
 export const convertToGraphQLUser = (userFromDB: DbUser): User => {
     const user: User = {
@@ -41,6 +43,7 @@ export const convertToGraphQLFloor = (floorFromDB: DbFloor): Floor => {
 export const convertToGraphQlArea = (areaFromDB: DbArea): Area => {
     const area: Area = {
         ...areaFromDB,
+        shape: JSON.stringify(areaFromDB.shape),
         databaseId: areaFromDB.id,
         id: "floor" + areaFromDB.id.toString()
     }
